feat(modules): added guild module shadowBan for: 912672204905385997
