--- src/modules/Guild/placeholder/placeholder/placeholder.txt
@@ -1,1 +1,0 @@
-This is just a placeholder file.
